Tests for conversions

+++ TestDataStructures/comparemacros.h
@@ -4,6 +4,6 @@
 #include <QtTest/QtTest>
 
 #define COMPARE_SIZE(list, expected_size) QCOMPARE((int) (list).size(), (expected_size))
-#define COMPARE_INTS(actual, expected) QCOMPARE((int) (actual), (int) (expected))
+#define COMPARE_INTS(actual, expected) QCOMPARE((unsigned long int) (actual), (unsigned long int) (expected))
 
 #endif // COMPAREMACROS_H
